// grayjay-android/app/src/unstable/assets/sources/bilibili/src/types.ts
export type BiliBiliCommentContext = {
    /** the id of the content */
    readonly oid: string
    /** the parent/root comment id */
    readonly rpid: string
    /** the type of comment "33" for course episode comments and "1" for everything else */
    readonly type: "33" | "1" | "11"
}

export type RequestMetadata<X> = {
    request(builder: BatchBuilder): BatchBuilder
    process(http_response: BridgeHttpResponse<string>): X
}

export type Params = {
    readonly [key: string]: string
}
// ...
export type SubCommentsResponse = {
    readonly data: {
        readonly page: {
            readonly count: number
        }
        readonly replies: Comment[]
    }
}

export type Comment = {
    readonly rpid: number
    readonly ctime: number
    readonly like: number
    readonly rcount: number
    readonly member: {
        readonly mid: number
        readonly uname: string
        readonly avatar: string
    }
    readonly content: {
        readonly message: string
    }
    readonly replies: {
        readonly rpid: number
    }[]
}

export type CommentsResponse = {
    readonly data: {
        readonly cursor: {
            is_end: boolean
        }
        readonly replies: Comment[]
        readonly top: {
            readonly upper: Comment | null
        }
    }
    readonly code: 0
} | {
    readonly code: -404
}
// ...
export type SeasonResponse = {
    readonly result: {
        readonly evaluate: string
        readonly episodes: {
            readonly ep_id: number
            readonly aid: number
            readonly bvid: string
            readonly cid: number
            readonly cover: string
            readonly duration: number
            readonly long_title: string
            readonly pub_time: number
        }[]
        readonly cover: string
        readonly title: string
        readonly stat: {
            readonly views: number
        }
    }
}
// ...
export type Card = {
    readonly item?: {
        readonly content?: string
        readonly description?: string
    }
    readonly dynamic?: string
}
// ...
export type VideoDetailResponse = {

    readonly code: number

    readonly data: {
        readonly View: {
            readonly title: string
            readonly pic: string
            readonly pubdate: number
            readonly cid: number
            readonly aid: number
            readonly videos: number
            readonly pages: {
                readonly cid: number
                readonly page: number
                readonly part: string
                readonly duration: number
                readonly first_frame: string
            }[]
            readonly owner: {
                readonly mid: number
                readonly name: string
                readonly face: string
            }
            readonly stat: {
                readonly view: number
                readonly like: number
            }
            readonly desc_v2: null | {
                readonly raw_text: string
            }[]
            readonly ugc_season?: {
                readonly id: number
                readonly title: string
                readonly cover: string
                readonly mid: number
                readonly sections: {
                    readonly season_id: number
                    readonly id: number
                    readonly title: string
                    readonly episodes: {
                        readonly id: number
                        readonly aid: number
                        readonly cid: number
                        readonly bvid: string
                        readonly title: string
                        readonly arc: {
                            readonly pic: string
                            readonly pubdate: number
                            readonly duration: number
                            readonly stat: {
                                readonly view: number
                            }
                        }
                        readonly page?: {
                            readonly cid: number
                            readonly part: string
                            readonly duration: number
                            readonly first_frame?: string
                        }
                    }[]
                }[]
            }
        }
        readonly Card: {
            readonly card: {
                readonly fans: number
            }
        }
        readonly Related: {
            readonly title: string
            readonly pic: string
            readonly pubdate: number
            readonly cid: number
            readonly aid: number
            readonly bvid: string
            readonly owner: {
                readonly mid: number
                readonly name: string
                readonly face: string
            }
            readonly stat: {
                readonly view: number
                readonly like: number
            }
            readonly duration: number
        }[]
    }
}